Map trace service buffers only when used.
This CL prevents the call-trace service from mapping all allocated buffers at buffer creation time, preferring to leave them as unmapped files when not in use. This causes them to be mapped into the call-trace service memory only when a buffer is being written to (process ended events) or when read from (when writing the buffer contents to disk). This should reduce contention on the VM address space in the call-trace service process.

BUG=
R=siggi@chromium.org

Review URL: https://codereview.appspot.com/10918043