Create dummy unmapped blocks for COFF BSS sections
This changes the representation of BSS sections from no blocks (pointers
to the BSS would go to the symbol table, as they do for external
symbols) to unmapped blocks: blocks with a size but no data and no
address in the original image layout.

This makes it easier to keep track of references, in anticipation of
COFF file writing support. It also matches more closely the way
uninitialized data is handled in PE files. All of that at the cost of
some added complexity in handling unmapped blocks in COFF files.

R=chrisha@chromium.org
BUG=

Review URL: https://codereview.appspot.com/12955043