Add testing_lib to Syzygy.
Initially, this contains functionality for allowing reflective invocation of the toolchain. The paths to the pieces of the toolchain are extracted from the build system by a Python generator, and exposed to the unittest code as a set of constants. Intended use is as follows:

TEST(Foo, Bar) {
  ASSERT_NO_FATAL_FAILURE(testing::SetToolchainPaths());
  base::CommandLine cmd_line(testing::kLinkerPath);
  ...
  base::LaunchProcess(cmd_line, ...);
}

BUG=
R=rogerm@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/13723044