Fix a DCHECK failed when relinking with basic block reordering.
Roger could you look over my shoulder on this one.
I think this is the right fix, but I don't get when this was introduced and why we never catch it before.

I plan to add a end2end for basic block re-ordering this week.

R=chrisha@chromium.org, rogerm@chromium.org

BUG=

Review URL: https://codereview.appspot.com/13842046